backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
bcrypt
PyJWT
pandas
python-dotenv
psycopg2-binary
